return long long

+++ ssl/ssl_sess.c
@@ -916,7 +916,7 @@ long SSL_SESSION_get_time(const SSL_SESSION *s)
     return (long)ossl_time_to_time_t(s->time);
 }
 
-long SSL_SESSION_get_rtt(const SSL_SESSION *s) // might have to return a different data type. Or pass in a uint64_t* from nginx to fill in
+long long SSL_SESSION_get_rtt(const SSL_SESSION *s) // might have to return a different data type. Or pass in a uint64_t* from nginx to fill in
 {
     if (s == NULL)
         return 0;
